+++ index.html
@@ -24,11 +24,11 @@
     <button class="awake">awake</button>
     <div class="speech-log">There would be your phrase</div>
     <div class="battery-charging"></div>
-    <audio controls>
+    <audio controls loop autoplay>
         <source src="./hru.mp3" type="audio/mp3">
     </audio>
-    <svg id="svg"></svg>
-    <script src="./index.js"></script>
+    <svg class="svg"></svg>
+    <script src="./main.js"></script>
 </body>
 
 </html>
